Fix integer overflow on transpose of big data
Some data sets can be very large (such as 1501x1792x2048). For big
data sets, an integer overflow would occur during the transpose from
C to Fortran or back, and it would cause the data to look very bad.

Cast the index to a size_t before the final multiplication to prevent
the integer overflow.

+++ tomviz/GenericHDF5Format.cxx
@@ -38,7 +38,8 @@ void ReorderArrayC(T* in, T* out, int dim[3])
   for (int i = 0; i < dim[0]; ++i) {
     for (int j = 0; j < dim[1]; ++j) {
       for (int k = 0; k < dim[2]; ++k) {
-        out[(i * dim[1] + j) * dim[2] + k] = in[(k * dim[1] + j) * dim[0] + i];
+        out[static_cast<size_t>(i * dim[1] + j) * dim[2] + k] =
+          in[static_cast<size_t>(k * dim[1] + j) * dim[0] + i];
       }
     }
   }
@@ -50,7 +51,8 @@ void ReorderArrayF(T* in, T* out, int dim[3])
   for (int i = 0; i < dim[0]; ++i) {
     for (int j = 0; j < dim[1]; ++j) {
       for (int k = 0; k < dim[2]; ++k) {
-        out[(k * dim[1] + j) * dim[0] + i] = in[(i * dim[1] + j) * dim[2] + k];
+        out[static_cast<size_t>(k * dim[1] + j) * dim[0] + i] =
+          in[static_cast<size_t>(i * dim[1] + j) * dim[2] + k];
       }
     }
   }
